Initial commit for MetaSoccer Universe (MSU) token
This is just a recular ERC20 token based on the great OpenZeppelin
contracts v4.1.

This token has:
AccessControl: MINT And PAUSER role
Pausable: as a red button emergency
Burnable: This is a requirement for our TBC, which we will be publishing
in the incoming days.

Other than that, this should be a standard and regular token.